requirements.txt
cassandra-driver==3.19.0
certifi==2019.9.11
chardet==3.0.4
Django==2.2.6
django-cassandra-engine==1.5.5
idna==2.8
pytz==2019.3
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.7
